adding browser-specific entry point for 6to5
