+++ package.json
@@ -2,6 +2,7 @@
   "name": "sprintly-data",
   "version": "0.0.5",
   "main": "index.js",
+  "browser": "sprintly-data.js",
   "repository": {
     "type": "git",
     "url": "http://github.com/sprintly/sprintly-data.git"
